requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-limiter==3.5.0
python-dotenv==1.0.0
pymongo==4.6.1
waitress==3.0.0
gunicorn==21.2.0
pytest==7.4.3
requests==2.31.0 